third_party/WebKit/LayoutTests/fast/css/fontface-constructor-error-expected.txt

Test FontFace constructor with invalid parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS faceWithWrongSrc.status is "error"
PASS rejectionValue.name is "SyntaxError"
PASS faceWithWrongDescriptors.status is "error"
PASS faceWithWrongDescriptors.style is "normal"
PASS faceWithWrongDescriptors.weight is "normal"
PASS faceWithWrongDescriptors.unicodeRange is "U+0-10FFFF"
PASS faceWithWrongDescriptors.variant is "normal"
PASS faceWithWrongDescriptors.featureSettings is "normal"
PASS rejectionValue.name is "SyntaxError"
PASS successfullyParsed is true

TEST COMPLETE
